+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 #### Changed
 - default value of boolean columns is set to false if it is not null (this prevent from errors when user forgot set default to false)
 
+#### Added
+- set autoincrement for new table and also for existing table
+
 ### [1.3.0] - 2020-04-16
 #### Added
 - method renameColumn() to migrations